# refactor(utils): optimize in-place array filtering ~70,000x faster