@qoo1476 2025-01-15
wow it is super fast 🚀🚀

but the thing is result is weired on my use case

you can copy it and just paste it on F12 console window and see the result

```
function filter(data, predicate) {
  return data.filter(predicate)
}

function filterInPlace(data, predicate) {
  for (let i = data.length; i--; i>=0) {
    if (!predicate(data[i], i, data))
      data.splice(i, 1)
  }
  return data
}

function turboFilter(data, predicate) {
  for (let i = data.length; i--; i >= 0) {
    if (!predicate(data[i], i, data)) {
      const last = --data.length;
      if (i < last) data[i] = data[last];
    }
  }
  return data;
}

console.log(filter([1, 2, 3, 4, 5, 6, 7, 8, 9], (v) => v % 3 === 0))
console.log(filterInPlace([1, 2, 3, 4, 5, 6, 7, 8, 9], (v) => v % 3 === 0))
console.log(turboFilter([1, 2, 3, 4, 5, 6, 7, 8, 9], (v) => v % 3 === 0))
```

the turboFilter result is [undefined, undefined, 3]

am i missing something?