@danielroe 2025-01-24
closing for now - but feel free to reopen any time with performance improvements!